Speed up of visual feedback with multiple frames
By default, in win-switch-(on|off)-feedback, the mode-line
background and foreground face are set. When multiple frames
are present, these functions make the change in all frames,
even though the mode-line face only applies to the active
mode line in the selected frame. This turns out to be
significantly slower (3-5 times) which becomes a notable
lag as more frames are used. By passing the selected
frame, this cost is eliminated and responsiveness is good.
Additional checks to ensure that the face is not changed
unnecessarily also seem to provide a useful speedup.

Also the pushback of keys to unread-command-events
in win-switch-exit-by-timeout was changed to be
consistent with the streamlined version made
in win-switch-exit-and-redo from the last commit.

Tests passed in GNU Emacs 25.1.1.