Fixed feedback bug for caused by recent speed-up
Changing the mode-line feedback to set for each frame eliminated a
noticeable and annoying slow-down when multiple frames are present. But
it left the possibility of the mode-line colors not being properly reset
after exit from window-switching mode. That has been fixed here.

A refactoring of the visual feedback mechanism is needed, as detailed in
comments within (see ATTN: Issue). That will be handled next, but this
commit fixes the problem while maintaining the good responsiveness.